+++ src/components/Hangman/index.tsx
@@ -1,2 +1,3 @@
 export * from "./AnswerBorad";
+export * from "./AnswerInput";
 export * from "./Heart";
